New base class with common platform functionality.
