+++ library/Zend/Db/Adapter/Platform/IbmDb2.php
@@ -9,7 +9,7 @@
 
 namespace Zend\Db\Adapter\Platform;
 
-class IbmDb2 implements PlatformInterface
+class IbmDb2 extends AbstractSql92BasedPlatform
 {
     protected $quoteValueAllowed = false;
 
@@ -51,16 +51,6 @@ class IbmDb2 implements PlatformInterface
     }
 
     /**
-     * Get quote indentifier symbol
-     *
-     * @return string
-     */
-    public function getQuoteIdentifierSymbol()
-    {
-        return '"';
-    }
-
-    /**
      * Quote identifier
      *
      * @param  string $identifier
@@ -71,7 +61,7 @@ class IbmDb2 implements PlatformInterface
         if ($this->quoteIdentifiers === false) {
             return $identifier;
         }
-        return '"' . str_replace('"', '\\' . '"', $identifier) . '"';
+        return parent::quoteIdentifier($identifier);
     }
 
     /**
@@ -90,16 +80,6 @@ class IbmDb2 implements PlatformInterface
             $identifierChain = implode('"' . $this->identifierSeparator . '"', $identifierChain);
         }
         return '"' . $identifierChain . '"';
-    }
-
-    /**
-     * Get quote value symbol
-     *
-     * @return string
-     */
-    public function getQuoteValueSymbol()
-    {
-        return '\'';
     }
 
     /**
@@ -137,25 +117,6 @@ class IbmDb2 implements PlatformInterface
     }
 
     /**
-     * Quote value list
-     *
-     * @param string|string[] $valueList
-     * @return string
-     */
-    public function quoteValueList($valueList)
-    {
-        if (!is_array($valueList)) {
-            return $this->quoteValue($valueList);
-        }
-
-        $value = reset($valueList);
-        do {
-            $valueList[key($valueList)] = $this->quoteValue($value);
-        } while ($value = next($valueList));
-        return implode(', ', $valueList);
-    }
-
-    /**
      * Get identifier separator
      *
      * @return string
@@ -177,31 +138,6 @@ class IbmDb2 implements PlatformInterface
         if ($this->quoteIdentifiers === false) {
             return $identifier;
         }
-        $parts = preg_split('#([\.\s\W])#', $identifier, -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY);
-
-        if ($safeWords) {
-            $safeWords = array_flip($safeWords);
-            $safeWords = array_change_key_case($safeWords, CASE_LOWER);
-        }
-        foreach ($parts as $i => $part) {
-            if ($safeWords && isset($safeWords[strtolower($part)])) {
-                continue;
-            }
-
-            switch ($part) {
-                case ' ':
-                case '.':
-                case '*':
-                case 'AS':
-                case 'As':
-                case 'aS':
-                case 'as':
-                    break;
-                default:
-                    $parts[$i] = '"' . str_replace('"', '\\' . '"', $part) . '"';
-            }
-        }
-
-        return implode('', $parts);
+        return parent::quoteIdentifierInFragment($identifier, $safeWords);
     }
 }
